requirements.txt
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
mongoengine==0.29.1
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
